requirements.txt
pyqt5
mss
pillow
imgcompare